refactor: reorganize code structure
