+++ docs/source/api/api.rst
@@ -134,7 +134,7 @@ Differentiable Meta-RMSProp Optimizer
 Implicit differentiation
 ========================
 
-.. currentmodule:: torchopt.implicit
+.. currentmodule:: torchopt.diff.implicit
 
 .. autosummary::
