+++ torchopt/linear_solve/normal_cg.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 # ==============================================================================
 # This file is modified from:
-# https://github.com/google/jaxopt/blob/main/jaxopt/_src/implicit_diff.py
+# https://github.com/google/jaxopt/blob/main/jaxopt/_src/linear_solve.py
 # ==============================================================================
 # Copyright 2021 Google LLC
 #
@@ -29,78 +29,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Linear algebra solvers."""
+"""Linear algebra solver for ``A^T A x = A^T b`` using conjugate gradient."""
 
 # pylint: disable=invalid-name
 
 import functools
 from typing import Callable, Optional
 
-import functorch
-
-from torchopt import linalg, pytree
+from torchopt import linalg
+from torchopt.linear_solve.utils import make_normal_matvec, make_ridge_matvec, make_rmatvec
 from torchopt.typing import TensorTree
 
 
-__all__ = ['solve_cg', 'solve_normal_cg']
-
-
-def tree_add(tree_x: TensorTree, tree_y: TensorTree, alpha: float = 1.0) -> TensorTree:
-    """Computes tree_x + alpha * tree_y."""
-    return pytree.tree_map(lambda x, y: x.add(y, alpha=alpha), tree_x, tree_y)
-
-
-def _make_rmatvec(
-    matvec: Callable[[TensorTree], TensorTree], x: TensorTree
-) -> Callable[[TensorTree], TensorTree]:
-    """Returns a function that computes A^T y from matvec(x) = A x."""
-    _, vjp, *_ = functorch.vjp(matvec, x)
-    return lambda y: vjp(y)[0]
-
-
-def _normal_matvec(matvec: Callable[[TensorTree], TensorTree], x: TensorTree) -> TensorTree:
-    """Computes A^T A x from matvec(x) = A x."""
-    matvec_x, vjp, *_ = functorch.vjp(matvec, x)
-    return vjp(matvec_x)[0]
-
-
-def _make_ridge_matvec(
-    matvec: Callable[[TensorTree], TensorTree], ridge: float = 0.0
-) -> Callable[[TensorTree], TensorTree]:
-    def ridge_matvec(v: TensorTree) -> TensorTree:
-        return tree_add(matvec(v), v, alpha=ridge)
-
-    return ridge_matvec
-
-
-def solve_cg(
-    matvec: Callable[['TensorTree'], 'TensorTree'],  # (x) -> A @ x
-    b: 'TensorTree',
-    ridge: Optional[float] = None,
-    init: Optional['TensorTree'] = None,
-    **kwargs,
-) -> 'TensorTree':
-    """Solves ``A x = b`` using conjugate gradient.
-
-    It assumes that ``A`` is a hermitian, positive definite matrix.
-
-    Args:
-        matvec: A function that returns the product between ``A`` and a vector.
-        b: A tree of tensors for the right hand side of the equation.
-        ridge: Optional ridge regularization.
-        init: Optional initialization to be used by conjugate gradient.
-        **kwargs: Additional keyword arguments for the conjugate gradient solver.
-
-    Returns:
-        The solution with the same structure as ``b``.
-    """
-    if ridge is not None:
-        #      (x) -> A @ x + ridge * x
-        # i.e. (x) -> (A + ridge * I) @ x
-        matvec = _make_ridge_matvec(matvec, ridge=ridge)
-
-    # Returns solution for `(A + ridge * I) @ x = b`.
-    return linalg.cg(matvec, b, x0=init, **kwargs)
+__all__ = ['solve_normal_cg']
 
 
 def _solve_normal_cg(
@@ -138,20 +79,18 @@ def _solve_normal_cg(
         # Returns solution for `A @ x = b`.
         return linalg.cg(matvec, b, x0=init, **kwargs)
 
-    rmatvec = _make_rmatvec(matvec, example_x)  # (x) -> A.T @ x
-
-    def normal_matvec(x):  # (x) -> A.T @ A @ x
-        return _normal_matvec(matvec, x)
+    rmatvec = make_rmatvec(matvec, example_x)  # (x) -> A.T @ x
+    normal_matvec = make_normal_matvec(matvec)  # (x) -> A.T @ A @ x
 
     if ridge is not None:
         #      (x) -> A.T @ A @ x + ridge * x
         # i.e. (x) -> (A.T @ A + ridge * I) @ x
-        normal_matvec = _make_ridge_matvec(normal_matvec, ridge=ridge)
+        normal_matvec = make_ridge_matvec(normal_matvec, ridge=ridge)
 
-    Ab = rmatvec(b)  # A.T @ b
+    rhs = rmatvec(b)  # A.T @ b
 
     # Returns solution for `(A.T @ A + ridge * I) @ x = A.T @ b`.
-    return linalg.cg(normal_matvec, Ab, x0=init, **kwargs)
+    return linalg.cg(normal_matvec, rhs, x0=init, **kwargs)
 
 
 def solve_normal_cg(**kwargs):
