+++ torchopt/schedule/polynomial.py
@@ -29,7 +29,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Learning rate schedules."""
+"""Polynomial learning rate schedules."""
 
 import logging
